requirements.txt
fastapi>=0.68.0
streamlit>=1.0.0
openai>=1.0.0
python-dotenv>=0.19.0
langchain>=0.0.200
langgraph>=0.0.5
python-multipart>=0.0.5
redis>=4.3.4
python-docx>=0.8.11
ffmpeg-python>=0.2.0
pydub>=0.25.1
